refactor: change the way regexes are handles
In the old situation we had to calculate whitespace, leading characters, parameters
Now we match the parameter group, giving us almost all the info we need.

We no longer need the utilities, should also make it easier to add new tags.

Note we assume that the line contents doesnt change for a SingleTag,
otherwise it will break the current implementation